python-backend/requirements.txt
fastapi[all]
uvicorn
pandas
scikit-learn
SQLAlchemy
databases[sqlite]
python-dotenv
